// src/core/enums/css-vars-enum.ts
/**
 * Constants instead of typescript enums ---> smaller bundle size
 */
export const AnimateOnClick = '--animate-onclick';
export const SliderWidth = '--width';
export const SliderHeight = '--height';
export const SliderRadius = '--panel-bg-border-radius';

export const SliderBg = '--panel-bg';
export const SliderBgHover = '--panel-bg-hover';
export const SliderBgFill = '--panel-bg-fill';

export const PointerWidth = '--pointer-width';
export const PointerHeight = '--pointer-height';
export const PointerRadius = '--pointer-border-radius';
export const PointerBg = '--pointer-bg';
export const PointerBgHover = '--pointer-bg-hover';
export const PointerBgFocus = '--pointer-bg-focus';
export const PointerShadow = '--pointer-shadow';
export const PointerShadowHover = '--pointer-shadow-hover';
export const PointerShadowFocus = '--pointer-shadow-focus';
export const PointerBorder = '--pointer-border';
export const PointerBorderHover = '--pointer-border-hover';
export const PointerBorderFocus = '--pointer-border-focus';

/*
export const enum CSSVariables {

  AnimateOnClick = '--animate-onclick',
  SliderWidth = '--width',
  SliderHeight = '--height',
  SliderRadius = '--panel-bg-border-radius',
  
  SliderBg = '--panel-bg',
  SliderBgHover = '--panel-bg-hover',
  SliderBgFill = '--panel-bg-fill',
  
  PointerWidth = '--pointer-width',
  PointerHeight = '--pointer-height',
  PointerRadius = '--pointer-border-radius',
  PointerBg = '--pointer-bg',
  PointerBgHover = '--pointer-bg-hover',
  PointerBgFocus = '--pointer-bg-focus',
  PointerShadow = '--pointer-shadow',
  PointerShadowHover = '--pointer-shadow-hover',
  PointerShadowFocus = '--pointer-shadow-focus',
  PointerBorder = '--pointer-border',
  PointerBorderHover = '--pointer-border-hover',
  PointerBorderFocus = '--pointer-border-focus',
}*/